[FIX] account_reports: allow printing expressions with "boolean" figure type
When a report contained an expression using "boolean" figure_type (for example, the annual LU tax report), trying to print it to pdf raised an error when rendering the qweb template, since the formatted 'name' key for that column contained a value of type bool instead of str ; and the template tried to call len(bool) at some point.

To fix that, we make it so that _format_value now directly returns 'Yes' and 'No'. The owl template of the report also needed to be fixed in order to use that value instead of hardcoding those two strings.

closes odoo/enterprise#50828

Signed-off-by: John Laterre (jol) <jol@odoo.com>