[FIX] account_avatax: neuter
This commit adds a "missing" part of the neutralisation of the
account_avatax app. This brings it back in line with the complete
neutralisation script that is used on the PaaS and SaaS in versions < 16.0

In a purely standard environment, this is not strictly necessary, as
with is_avatax=false, we should not even read avalara_environment
anymore (except for some basic actions) , but the presence of
a settings that still potentially explicitly says "production"
when the database is properly neutralised causes confussion to the end
users.

It is far from ideal to have fields that do not reflect the exact
reality of the behaviour of the app, but changing that is not an option
in stable, nor is appropriate the scope of this fix.

opw-3593983

closes odoo/enterprise#50889

Signed-off-by: William André (wan) <wan@odoo.com>