[FIX] sign: ignore archived signature requests
Issue:
------
The purpose of the `signature_count` field is to count the number of signature requests.
However, it counts signatures that have been archived.
For example, in the recruitment application, when we want to see the number of requests
sent to a job application.

Solution:
---------
Modify the domain to take status into account and display only signatures that are `sent` and `completed`.

opw-3550893

closes odoo/enterprise#50907

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>