[FIX] spreadsheet_edition: allow to replace attachment
Business analysts wants to transfert spreadsheet json file from one database
to another.
They would prepare the spreadsheet in a staging/dev environnment, then download
the json file and upload it to the customer database.

Fiduciaries also want to do this. They want to copy spreadsheets to all of their
customer databases.

There's currently no real fonctional way to do it... but there's a hack:
Create a new dummy spreadsheet document, then replace the attachment file which
underlies the document.
This hack however only works on version 16.4+

The reason it doesn't work here is because as soon as the dummy spreadsheet is
open, the field `spreadsheet_snapshot` is filled with the initial data from
the attachments. The next time the spreadsheet is open, the data comes from
the snapshot, and not the attachment. So replacing it has not effect.

This commit backports a change made in saas-16.4

closes odoo/enterprise#50917

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>