[FIX] account_reports: use correct cog button target on composite reports
In debug mode, a little cog button is displayed next to the report filters. Clicking it open the form view of the report.

The problem is the report used as the target of this action was options['report_id'], not options['sections_source_id']. Because of that, when clicking this button on a composite report, only the form view of the currently displayed section was open, while we actually should have opened the form of the whole composite report.

Part-of: odoo/enterprise#50927