[FIX] account_reports: don't depend on show_debug_column options to display the cog button
The cog button allows opening the form view of the report to reconfigure it. It is only supposed to be displayed in debug mode. For that, the show_debug_column option key was used. Yet, this was wrong, as this option key is not always true in debug mode: for example, if a comparison is being done, it will be False. We now display the cog directly from debug mode info, without depending on options.

closes odoo/enterprise#50927

Signed-off-by: John Laterre (jol) <jol@odoo.com>