[FIX] project_forecast: freeze time for test_shift_creation_from_project
Before this commit, the `test_shift_creation_from_project` test failed
during the weekend because the working schedule of the current company
considers the weekend is not a working day and so instead of creating a
planning template during 2 days, it will create a planning template during
one day.

This commit freezes the time to be sure, the test is executed during a
working day to stay valid and keep the same behavior to test.

runobt-41215

X-original-commit: 013e5fcca447fe8f6fd217617af626709eaa5213
Part-of: odoo/enterprise#51068