[FIX] web_enterprise: fix tests failing randomly
Commit [1] adapted a few WebClientEnterprise tests to reflect with
a community change [2] which makes a click a on menu perform two
updates of the webclient: one with a blank screen, and another one
with the requested view.

Tests were adapted to check that after a tick, no view was
displayed yet, and that after another tick, the view was finally
there. This is incorrect and sometimes fails. Indeed, it might
happen that the kanban view is already there after the first tick,
because more than one animation frame can occur during a nextTick.

This commit fixes those test by simply removing the first assertion
(blank screen), which isn't the concern of those tests anyway.

[1] 430b9dc15031bce98e3be55b23438ee09494dada
[2] https://github.com/odoo/odoo/pull/124068

Task 3603040

closes odoo/enterprise#51091

X-original-commit: 0f2135238c6b0729ead6ec59560624e07017b913
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>