[FIX] pos_blackbox_be: fix wrong override in product screen
Prior to this commit, the override made on the method _setValue in
the product screen was including a call to the parent method without
the correct arguments leading to no behaviour when clicking on the
numpad in the product screen when the pos_blackbox_be module was
installed (e.g. when trying to do a discount). This commit fixes this
error.

closes odoo/enterprise#51111

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>