[FIX] account_reports: Fix report line label in report form view
Before this commit we use column.name as label of our columns
for report lines in the report form view.
It caused that the label was in lower case.
The aim of this commit is using the label instead of the name to
have the column label well formatted.

no task id

closes odoo/enterprise#51057

Signed-off-by: John Laterre (jol) <jol@odoo.com>